sha-crypt: Change parameter type and make hashing infallible (#500)
* sha-crypt: Change `rounds` parameter type to `u32`

Its type doesn't need to depend on the target architecture pointer
width.

Fixes #498.

* sha-crypt: Make hashing infallible

The rounds were already checked to be sane in
`Sha{256,512}Parameters::new`.

Fixes #499.

+++ sha-crypt/src/params.rs
@@ -4,18 +4,18 @@ use crate::errors;
 use core::default::Default;
 
 /// Default number of rounds.
-pub const ROUNDS_DEFAULT: usize = 5_000;
+pub const ROUNDS_DEFAULT: u32 = 5_000;
 
 /// Minimum number of rounds allowed.
-pub const ROUNDS_MIN: usize = 1_000;
+pub const ROUNDS_MIN: u32 = 1_000;
 
 /// Maximum number of rounds allowed.
-pub const ROUNDS_MAX: usize = 999_999_999;
+pub const ROUNDS_MAX: u32 = 999_999_999;
 
 /// Algorithm parameters.
 #[derive(Debug, Clone)]
 pub struct Sha512Params {
-    pub(crate) rounds: usize,
+    pub(crate) rounds: u32,
 }
 
 impl Default for Sha512Params {
@@ -28,7 +28,7 @@ impl Default for Sha512Params {
 
 impl Sha512Params {
     /// Create new algorithm parameters.
-    pub fn new(rounds: usize) -> Result<Sha512Params, errors::CryptError> {
+    pub fn new(rounds: u32) -> Result<Sha512Params, errors::CryptError> {
         if (ROUNDS_MIN..=ROUNDS_MAX).contains(&rounds) {
             Ok(Sha512Params { rounds })
         } else {
@@ -40,7 +40,7 @@ impl Sha512Params {
 /// Algorithm parameters.
 #[derive(Debug, Clone)]
 pub struct Sha256Params {
-    pub(crate) rounds: usize,
+    pub(crate) rounds: u32,
 }
 
 impl Default for Sha256Params {
@@ -53,7 +53,7 @@ impl Default for Sha256Params {
 
 impl Sha256Params {
     /// Create new algorithm parameters.
-    pub fn new(rounds: usize) -> Result<Sha256Params, errors::CryptError> {
+    pub fn new(rounds: u32) -> Result<Sha256Params, errors::CryptError> {
         if (ROUNDS_MIN..=ROUNDS_MAX).contains(&rounds) {
             Ok(Sha256Params { rounds })
         } else {
